Improve slice/index inlining logic
We generally can't inline into slice/index nodes because we can't have
expressions such as `(x + y)[1:0]`.  However, we can have nested
index/slice nodes, so we improve the blacklisting logic to not block
inlining if we see a slice or index driver.

Note, we assume that the index/slice driver is well formed, so if we
have a driver of the form <...>[3:0], the contents of the slice (<...>)
are assumed to be valid (so it could contain an ID, another slice, or an
index).